requirements.txt
click==8.1.7
colorama==0.4.6
filetype==1.2.0
Pillow==10.1.0
python-magic-bin==0.4.14
